Updated Movement
Updated movement to use fixed point math.
Created camera object.
drawGrid method implemented.

Not working yet. All black.

+++ MYBL_AB/MYBL_AB.ino
@@ -57,7 +57,7 @@ void loop() {
   if (soundYesNo == true) arduboy.audio.on();
   else arduboy.audio.off();
   //arduboy.clearDisplay();
-  arduboy.fillScreen(gameState == STATE_GAME_PLAYING);
+  arduboy.fillScreen((byte)(gameState == STATE_GAME_PLAYING));
   mainGameLoop[gameState]();
   arduboy.display();
 }
